Use Long for flex balance total and remove toIntExact overflow risk

+++ src/main/java/com/example/flexapp/dto/FlexBalanceResponse.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 public class FlexBalanceResponse {
 
     private Long userId;
-    private Integer totalFlexMinutes;
+    private Long totalFlexMinutes;
 }
